requirements.txt
# Flask and extensions
flask>=3.0.0
flask-sqlalchemy>=3.1.0
flask-login>=0.6.3
flask-wtf>=1.2.0
flask-migrate>=4.0.0

# Database
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.0

# Utilities
python-dotenv>=1.0.0
gunicorn>=21.0.0

# Development
pytest>=8.0.0
